ユニークレートを0に設定していてもユニークアイテム（名称がHQとなるが性能はノーマル）をドロップする場合がある。 fixes #469
git-svn-id: http://svn.l1j.org/l1j/trunk@424 698199e4-35c4-443b-82c4-7f4b4ae73c13

+++ src/jp/l1j/server/datatables/DropTable.java
@@ -178,7 +178,7 @@ public class DropTable {
 			// ユニークオプションを付加
 			double uniqueRateOfItem = DropRateTable.getInstance().getUniqueRate(itemId);
 			double uniqueRate = uniqueDropRate * uniqueRateOfMapId * uniqueRateOfItem;
-			item.setUniqueOptions(uniqueRate);
+			item.setUniqueOptions((int) uniqueRate);
 		}
 	}
